backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
starlette==0.38.6
python-multipart==0.0.9
pydantic==2.9.2
redis==5.0.8
langchain==0.3.0
langchain-community==0.3.0
langchain-ollama==0.3.2
pypdf==4.3.1
python-docx==1.1.2
prometheus-client==0.20.0
psutil==6.0.0
pydantic-settings==2.5.2
chromadb==0.5.11
langchain-chroma==0.1.4